+++ src/App.css
@@ -8,4 +8,5 @@
 
 body {
   font-family: "Nunito Sans", sans-serif;
+  background: #f2f2f2;
 }
