+++ src/components/Countries/Countries.module.css
@@ -1,0 +1,54 @@
+.countries {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 81px;
+  margin: 60px;
+  justify-content: center;
+  align-items: center;
+}
+
+.country {
+  height: 336px;
+  width: 264px;
+  background: #ffffff;
+  box-shadow: 0px 0px 7px 2px rgba(0, 0, 0, 0.0294384);
+  border-radius: 5px;
+}
+
+.flag {
+  width: 100%;
+  height: 160px;
+  object-fit: cover;
+}
+
+.info {
+  margin-left: 24px;
+}
+
+.name {
+  font-style: normal;
+  font-weight: 800;
+  font-size: 18px;
+  line-height: 26px;
+  color: #111517;
+  margin: 24px 0px;
+}
+
+h4 {
+  font-family: "Nunito Sans";
+  font-style: normal;
+  font-weight: 800;
+  font-size: 14px;
+  line-height: 16px;
+  color: #111517;
+  margin: 5px 0px;
+}
+
+span {
+  font-family: "Nunito Sans";
+  font-style: normal;
+  font-weight: 600;
+  font-size: 14px;
+  line-height: 16px;
+  color: #111517;
+}
