+++ src/components/Header/Header.module.css
@@ -6,4 +6,5 @@
   justify-content: space-between;
   align-items: center;
   padding: 0px 80px;
+  background-color: #ffffff;
 }
