+++ src/components/Input/Input.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Input = () => {
+  return (
+    <div>
+      <input type="text" />
+    </div>
+  );
+};
+
+export default Input;
