+++ src/pages/Home.jsx
@@ -1,10 +1,12 @@
 import React from "react";
+import Countries from "../components/Countries/Countries";
 import Header from "../components/Header/Header";
 
 const Home = () => {
   return (
     <div>
       <Header />
+      <Countries />
     </div>
   );
 };
